10. Create a new user role.
Create a new user role for a given user id, unit id, role id, an optional valid from timestamp (if not specified, default to the current date and time) and an optional valid to timestamp (if not specified, default to no timestamp). If a valid to timestamp is specified, it must be after the valid from timestamp (or the current date and time if valid from timestamp is not specified in the request). At most one user role for a given combination of user id, unit id and role id can be valid at any point in time.