11. Update an existing user role.
Only the valid from and valid to timestamps can be changed. The valid from timestamp, if specified, must be a timestamp (a user role must always have a valid from timestamp). The requirement that the valid to timestamp, if specified, must come after the valid from timestamp must be enforced, and an update that would cause two user roles for the same user id, unit id and role id to be valid at the same time must be rejected.